refactor(ws_v2): ws_v2.subscriptions was renamed to ws_v2.ws_subscriptions for consistency
