+++ ibind/subscriptions/__init__.py
@@ -1,6 +1,6 @@
 from ibkr_ws_v2.ibkr_subscriptions import MarketDataSubscription, OrdersSubscription, AccountLedgerSubscription, AccountSummarySubscription, PnlSubscription, TradesSubscription, MarketHistorySubscription
 
-from ws_v2.subscriptions import SubscriptionHandle, BindingStatus, Subscription, SubscriptionResolver
+from ws_v2.ws_subscriptions import SubscriptionHandle, BindingStatus, Subscription, SubscriptionResolver
 
 __all__ = [
     'Subscription',
